backend/requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
argcomplete==1.10.3
beautifulsoup4==4.8.2
black==21.4b2
boto3==1.26.79
botocore==1.29.79
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
compressed-rtf==1.0.6
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@82a57ce0b70057685962b352535147d9a8118578
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
filelock==3.9.1
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.39.0
future==0.18.3
fvcore==0.1.5.post20221221
google-api-core==1.34.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-cloud-vision==1.0.0
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.48.2
hydra-core==1.3.2
idna==3.4
IMAPClient==2.1.0
iopath==0.1.9
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
layoutparser==0.3.4
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
pdf2image==1.16.2
pdfminer.six==20221105
pdfplumber==0.8.0
Pillow==9.4.0
portalocker==2.7.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
pycryptodome==3.17
pydot==1.4.2
pyparsing==3.0.9
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.1
six==1.12.0
sklearn==0.0.post1
sortedcontainers==2.4.0
soupsieve==2.4
SpeechRecognition==3.8.1
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
textract==1.6.5
threadpoolctl==3.1.0
toml==0.10.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
Wand==0.6.11
Werkzeug==2.2.2
xlrd==1.2.0
XlsxWriter==3.0.8
yacs==0.1.8
